Use piv-go/v2 and support ed25519 keys in agent
YubiKey 5.7 supports ed25519 keys, and these are fully spported
agent-wise with this commit. A fork of piv-go is used to allow this to
work, but this will probably not be needed for long.

The setup part of yubikey-agent has just been adjusted to work with
piv-go/v2, and still generates ecdsa keys.

Signed-off-by: Philip Eklöf <phi@pxy.se>